refactor: define SegmentEncoder type
