+++ Sources/AWSXRayRecorder/Encoder.swift
@@ -1,0 +1,17 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the aws-xray-sdk-swift open source project
+//
+// Copyright (c) 2020 pokryfka and the aws-xray-sdk-swift project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+extension XRayRecorder {
+    /// Encodes segment to string with JSON document.
+    public typealias SegmentEncoder = (XRayRecorder.Segment) throws -> String
+}
