+++ Sources/AWSXRayTesting/FoundationJSON.swift
@@ -1,0 +1,34 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the aws-xray-sdk-swift open source project
+//
+// Copyright (c) 2020 pokryfka and the aws-xray-sdk-swift project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+import AWSXRayRecorder
+import struct Foundation.Data
+import class Foundation.JSONEncoder
+
+private extension JSONEncoder {
+    func encode<T: Encodable>(_ value: T) throws -> String {
+        String(decoding: try encode(value), as: UTF8.self)
+    }
+}
+
+private let jsonEncoder: JSONEncoder = {
+    let encoder = JSONEncoder()
+    encoder.outputFormatting = .prettyPrinted
+    return encoder
+}()
+
+public enum FoundationJSON {
+    public static let segmentEncoder: XRayRecorder.SegmentEncoder = { segment in
+        try jsonEncoder.encode(segment)
+    }
+}
